Move fixtures closer to the tests that use them
Only `Counter` remains truly general since it's used in ABI and AST tests.
