+++ crates/solidity-v2/outputs/cargo/slang_solidity/src/tests/abi/mod.rs
@@ -4,3 +4,39 @@ mod selectors;
 mod storage_layout;
 
 use super::fixtures;
+use crate::define_fixture;
+
+define_fixture!(
+    AbiWithTuples,
+    file: "main.sol", r#"
+contract Test {
+    struct S { uint a; uint[] b; T[] c; }
+    struct T { uint x; uint y; }
+    function f(S memory, T memory, uint) public pure {}
+    function g() public pure returns (S memory, T memory, uint) {}
+}
+"#,
+);
+
+define_fixture!(
+    FullAbi,
+    file: "main.sol", r#"
+// SPDX-License-Identifier: UNLICENSED
+pragma solidity ^0.8;
+
+interface Base {
+    error SomethingWrong(string);
+    event BaseEvent(uint a, string m) anonymous;
+}
+
+contract Test is Base {
+    bytes32 public b;
+    constructor() { b = hex"12345678901234567890123456789012"; }
+    event Event(uint indexed a, bytes32 b);
+    error InsufficientBalance(uint256 available, uint256 required);
+    function foo(uint a) public { emit Event(a, b); }
+    receive() external payable { }
+    fallback() external { }
+}
+"#,
+);
