+++ crates/solidity-v2/outputs/cargo/slang_solidity/src/tests/ast/number_literals.rs
@@ -1,8 +1,47 @@
 use num_bigint::BigInt;
 use num_rational::BigRational;
 
-use super::fixtures;
 use crate::ast::{self, Number};
+use crate::define_fixture;
+
+define_fixture!(
+    NumberLiterals,
+    file: "main.sol", r#"
+// SPDX-License-Identifier: UNLICENSED
+pragma solidity ^0.8.29;
+
+contract NumberLiterals {
+    // Hex literals
+    uint constant HEX_SMALL = 0xff;
+    uint constant HEX_MIXED_CASE = 0xDeAdBeEf;
+    uint constant HEX_WITH_SEP = 0x1234_5678;
+    uint constant HEX_ZERO = 0x0;
+
+    // Address literal: a hex literal of exactly 40 digits is typed as an address.
+    address constant ADDRESS_BEEF = 0x000000000000000000000000000000000000beef;
+
+    // Decimal integer literals
+    uint constant DEC_ZERO = 0;
+    uint constant DEC_INT = 42;
+    uint constant DEC_WITH_SEP = 1_000_000;
+
+    // Decimal with scientific notation
+    uint constant DEC_EXP_POS = 1e3;
+    uint constant DEC_EXP_NEG_INT = 1500e-2;
+
+    // Decimal with units
+    uint constant ONE_WEI = 1 wei;
+    uint constant ONE_GWEI = 1 gwei;
+    uint constant ONE_ETHER = 1 ether;
+    uint constant HALF_ETHER = 0.5 ether;
+    uint constant ONE_HOUR = 1 hours;
+
+    // Rational decimal literals folded to integers through multiplication
+    uint constant FROM_HALF = 0.5 * 4;
+    uint constant FROM_ONE_AND_HALF = 1.5 * 2;
+}
+"#,
+);
 
 #[derive(Default)]
 struct NumberLiteralCollector {
@@ -39,7 +78,7 @@ fn rational(numerator: i64, denominator: u64) -> Number {
 
 #[test]
 fn test_hex_number_expression_value() {
-    let unit = fixtures::NumberLiterals::build_compilation_unit();
+    let unit = NumberLiterals::build_compilation_unit();
     let ast = unit.file("main.sol").unwrap().ast();
 
     let mut collector = NumberLiteralCollector::default();
@@ -73,7 +112,7 @@ fn test_hex_number_expression_value() {
 
 #[test]
 fn test_decimal_number_expression_value() {
-    let unit = fixtures::NumberLiterals::build_compilation_unit();
+    let unit = NumberLiterals::build_compilation_unit();
     let ast = unit.file("main.sol").unwrap().ast();
 
     let mut collector = NumberLiteralCollector::default();
